requirements.txt
# Kafka Python Client
kafka-python>=2.2.15

# Optional: For data visualization
# matplotlib==3.8.0
# pandas==2.1.0

